tests: delete and move around unit tests
