--- tests/app.test.js
@@ -1,56 +1,0 @@
-const http = require('node:http')
-const app = require('../src/app')
-
-const closeServer = async function (server) {
-  if (!server) {
-    return
-  }
-
-  await new Promise((resolve, reject) => {
-    server.close((error) => {
-      if (error) {
-        reject(error)
-        return
-      }
-
-      resolve()
-    })
-
-    if (typeof server.closeIdleConnections === 'function') {
-      server.closeIdleConnections()
-    }
-
-    if (typeof server.closeAllConnections === 'function') {
-      server.closeAllConnections()
-    }
-  })
-}
-// test express app deployment
-describe('app entrypoint', () => {
-  let baseUrl
-  let server
-
-  beforeAll(async () => {
-    server = http.createServer(app)
-
-    await new Promise((resolve) => {
-      server.listen(0, '127.0.0.1', () => {
-        baseUrl = `http://127.0.0.1:${server.address().port}`
-        resolve()
-      })
-    })
-  })
-
-  afterAll(async () => {
-    await closeServer(server)
-  })
-
-  test('redirects the root path to login', async () => {
-    const response = await fetch(`${baseUrl}/`, {
-      redirect: 'manual'
-    })
-
-    expect(response.status).toBe(302)
-    expect(response.headers.get('location')).toBe('/login')
-  })
-})
